+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - All nodes with `indentationLevel` attribute can be indented, instead of only hardcoded nodes.
+- Add a `setDocumentAttribute` on `SayController`
+- Export `SetDocAttributeStep`
 - Check if table can be inserted
 
 ### Fixed